fix: GameBoard: drawTileString moved to Tile TU
