+++ src/tile.cpp
@@ -1,5 +1,7 @@
 #include "tile.hpp"
 #include <cmath>
+#include <iomanip>
+#include <sstream>
 #include <vector>
 
 Color::Modifier Tile::tileColor(ull value) {
@@ -9,4 +11,15 @@ Color::Modifier Tile::tileColor(ull value) {
   int index = log < 12 ? log - 1 : 10;
 
   return colors[index];
+}
+
+std::string drawTileString(Tile currentTile) {
+  std::ostringstream tile_richtext;
+  if (!currentTile.value) {
+    tile_richtext << "    ";
+  } else {
+    tile_richtext << currentTile.tileColor(currentTile.value) << bold_on
+                  << std::setw(4) << currentTile.value << bold_off << def;
+  }
+  return tile_richtext.str();
 }
